bugfix; don't correct if depth>=5 (by default)
fermi2 corrects more errors than fermi1, but it is slower in the correction
phase. It is possible to take the fermi1 strategy to jump between kmers, but
this needs some code change. I will see how it performs on human data. All I am
doing is on E. coli only.